change hex to dec to 16 bits

+++ lib.h
@@ -21,7 +21,7 @@ typedef struct {
     int8_t nb_bytes;
 } instructions_details;
 
-int8_t hex_to_dec(char* str);
+int16_t hex_to_dec(char* str);
 void readFile(char* path);
 instructions_details* decodeInstructionArguments(int8_t first_byte, FILE* input, FILE* output);
 void decodeWhenInstructionNameLengthEqualsEight(int8_t first_byte, instructions_details* details, FILE* input, FILE* output);
